test: Remove checkLeaks test.
We've seen failures in Chrome 109 on Mojave where the Core#send beforeEach hook
has errored with:

    global leak detected: ret_nodes

    https://automate.browserstack.com/dashboard/v2/builds/d02d09eba240bfe63d5d4a38186f8b263e51448b/sessions/cf7edd2c7efa5ea1d0b8d2267290fa7bdcb0595a?auth_token=e9574e9467e1f22131acef157981aeb0791f3016cc7568e5faf4fbbf550f6402&build_token=aWYzOUl2OVQxRVBrWURhSUdKZWliZU1TV01NTStuQlNUejdRSTE1djgrWXNXWUVRZmh0d25KM3lSSHVKaC91NmFRRjRreFN6RnhMWFNsOS9pb25IM0E9PS0tejEwVU1hSlNsSHJKczZDK0Y3YjlHdz09--d2cf7eb2a908ea8939f86035ad68730cc7ea2969

I've tried to track the source of ret_nodes, but been unable to find it in our
code, our vendored code, or the test-framework dependencies in
test/unit/index.html. I think we're fine without the check.