CI: Switch from gsutil to rclone
The RavelinJS integration test takes around 3 minutes to run. Half of
this time is spent uploading the cipher. For example, see here, where
1m23s was spent uploading the cipher:

https://console.cloud.google.com/cloud-build/builds;region=global/520f0db3-c321-4fdd-9093-3d0269f1e65c;step=3?project=ravelin-builds

The file is around 750B in size. This is a net transfer rate of
approximately 9B/s.

In reality, the time is actually spent pulling
gcr.io/google.com/cloudsdktool/cloud-sdk:latest. As of the time of
writing, this is over 3GB in size.

This is idiotic.