CI: Try E2_HIGHCPU_32 as the machine type
The tests are faster, but still pretty slow; it's now spending most of
its time installing dependencies (~1m). This commit switches to a larger
machine type.

There's a risk that this'll lead to scheduling issues, as Cloud Build
may now need to wait until a machine of this type if available. There's
also a risk that this won't in fact speed things up. (Perhaps the
rate-determining step is downloading a huge number of things from
npm, but it's not like I can measure anything easily in CI.)